Fix issue with fetching all students (with School)
